Move environment related classes into the environment folder
