+++ src/worldofzuul/Game.java
@@ -1,5 +1,6 @@
 package worldofzuul;
 
+import worldofzuul.environment.Environment;
 import worldofzuul.character.Cook;
 import worldofzuul.character.Trump;
 import worldofzuul.character.Player;
